+++ src/modules/person/domain/person-time-limit-info.service.ts
@@ -1,10 +1,11 @@
 import { Injectable } from '@nestjs/common';
 import { Personenkontext } from '../../personenkontext/domain/personenkontext.js';
-import { PersonTimeLimitInfo } from '../domain/person-time-limit-info.js';
 import { DBiamPersonenkontextService } from '../../personenkontext/domain/dbiam-personenkontext.service.js';
-import { TimeLimitOccasion } from '../domain/time-limit-occasion.enums.js';
 import { PersonRepository } from '../../person/persistence/person.repository.js';
 import { Person } from '../../person/domain/person.js';
+import { PersonTimeLimitInfo } from './person-time-limit-info.js';
+import { TimeLimitOccasion } from './time-limit-occasion.enums.js';
+import { KOPERS_DEADLINE_IN_DAYS } from './person-time-limit.js';
 
 @Injectable()
 export default class PersonTimeLimitService {
@@ -27,7 +28,7 @@ export default class PersonTimeLimitService {
                 await this.dBiamPersonenkontextService.getKopersPersonenkontext(person.id);
             if (kopersKontext) {
                 const kopersdeadline: Date = new Date(kopersKontext.createdAt);
-                kopersdeadline.setDate(kopersdeadline.getDate() + 56);
+                kopersdeadline.setDate(kopersdeadline.getDate() + KOPERS_DEADLINE_IN_DAYS);
                 lockInfos.push(new PersonTimeLimitInfo(TimeLimitOccasion.KOPERS, kopersdeadline));
             }
         }
