+++ src/modules/person/domain/person-time-limit.ts
@@ -1,0 +1,1 @@
+export const KOPERS_DEADLINE_IN_DAYS: number = 56;
