+++ src/modules/person/domain/time-limit-occasion.enums.ts
